move actionChangeFontSize as well
